Refactor des popups à côté des layers
