+++ src/components/Map/components/layers/batimentsRaccordesReseauxChaleurFroid.tsx
@@ -1,6 +1,8 @@
+import Link from '@/components/ui/Link';
+import { formatMWhAn } from '@/utils/strings';
 import { type ExtractKeysOfType } from '@/utils/typescript';
 
-import { ifHoverElse, type MapSourceLayersSpecification } from './common';
+import { ifHoverElse, type PopupStyleHelpers, type MapSourceLayersSpecification } from './common';
 import { type MapConfiguration } from '../../map-configuration';
 import { type MapLayerSpecification } from '../../map-layers';
 
@@ -67,6 +69,7 @@ function buildLayerAndHoverLayer<LayerId extends string>(
         'icon-opacity': ['interpolate', ['linear'], ['zoom'], 9.2, 0, 10.5, ifHoverElse(0, batimentsRaccordesReseauxChaleurFroidOpacity)],
       },
       isVisible: (config) => config[conf.layerConfKey],
+      popup: Popup,
     },
     {
       id: `${conf.id}-hover`,
@@ -86,4 +89,40 @@ function buildLayerAndHoverLayer<LayerId extends string>(
       unselectable: true,
     },
   ] as const satisfies ReadonlyArray<MapLayerSpecification>;
+}
+
+const secteurBatimentRaccordeToLabels = {
+  A: 'Agriculture',
+  I: 'Industrie',
+  R: 'Résidentiel',
+  T: 'Tertiaire',
+};
+
+type BatimentRaccordeReseauxChaleurFroid = {
+  fid: number;
+  id_reseau: string;
+  filiere: 'C' | 'F';
+  adresse?: string;
+  code_grand_secteur: 'A' | 'I' | 'R' | 'T';
+  conso?: number;
+};
+
+function Popup(batimentRaccordeReseauxChaleurFroid: BatimentRaccordeReseauxChaleurFroid, { Property, Title }: PopupStyleHelpers) {
+  return (
+    <section>
+      {batimentRaccordeReseauxChaleurFroid.adresse && <Title>{batimentRaccordeReseauxChaleurFroid.adresse}</Title>}
+      <Property
+        label={`Consommation de ${batimentRaccordeReseauxChaleurFroid.filiere === 'C' ? 'chaleur' : 'froid'}`}
+        value={batimentRaccordeReseauxChaleurFroid.conso}
+        formatter={formatMWhAn}
+      />
+      <Property label="Secteur" value={secteurBatimentRaccordeToLabels[batimentRaccordeReseauxChaleurFroid.code_grand_secteur]} />
+      <strong>Identifiant du réseau&nbsp;:</strong>&nbsp;
+      <Link href={`/reseaux/${batimentRaccordeReseauxChaleurFroid.id_reseau}`} isExternal>
+        {batimentRaccordeReseauxChaleurFroid.id_reseau}
+      </Link>
+      <br />
+      <Property label="Source" value="SDES pour 2023" />
+    </section>
+  );
 }
