+++ src/components/Map/components/layers/communesFortPotentielPourCreationReseauxChaleur.tsx
@@ -1,6 +1,7 @@
 import { type Interval } from '@/utils/interval';
+import { formatMWhAn } from '@/utils/strings';
 
-import { ifHoverElse, type MapSourceLayersSpecification } from './common';
+import { ifHoverElse, type PopupStyleHelpers, type MapSourceLayersSpecification } from './common';
 
 export const communesFortPotentielPourCreationReseauxChaleurLayerColor = '#FF8329';
 export const communesFortPotentielPourCreationReseauxChaleurLayerOpacity = 0.7;
@@ -42,7 +43,42 @@ export const communesFortPotentielPourCreationReseauxChaleurLayersSpec = [
           ['<=', ['get', 'population'], config.communesFortPotentielPourCreationReseauxChaleur.population[1]],
         ],
         isVisible: (config) => config.communesFortPotentielPourCreationReseauxChaleur.show,
+        popup: Popup,
       },
     ],
   },
 ] as const satisfies ReadonlyArray<MapSourceLayersSpecification>;
+
+type CommuneFortPotentielPourCreationReseauxChaleur = {
+  id: number;
+  insee_dep: string;
+  insee_com: string;
+  nom: string;
+  population: number;
+  zones_fort_potentiel_total: number;
+  zones_fort_potentiel_chauf_mwh: number;
+  zones_fort_potentiel_ecs_mwh: number;
+};
+
+function Popup(
+  communeFortPotentielPourCreationReseauxChaleur: CommuneFortPotentielPourCreationReseauxChaleur,
+  { Property, Title }: PopupStyleHelpers
+) {
+  return (
+    <>
+      <Title>{communeFortPotentielPourCreationReseauxChaleur.nom}</Title>
+      <Property label="Nombre d'habitants" value={communeFortPotentielPourCreationReseauxChaleur.population} />
+      <Property
+        label="Besoins en chauffage sur les zones à fort potentiel (cumul)"
+        value={communeFortPotentielPourCreationReseauxChaleur.zones_fort_potentiel_chauf_mwh}
+        formatter={formatMWhAn}
+      />
+      <Property
+        label="Besoins en ECS sur les zones à fort potentiel (cumul)"
+        value={communeFortPotentielPourCreationReseauxChaleur.zones_fort_potentiel_ecs_mwh}
+        formatter={formatMWhAn}
+      />
+      <Property label="Source" value="BRGM" />
+    </>
+  );
+}
