+++ src/components/Map/components/layers/enrr-mobilisables/parkings.tsx
@@ -1,6 +1,6 @@
 import { darken } from '@/utils/color';
 
-import { ifHoverElse, type MapSourceLayersSpecification } from '../common';
+import { ifHoverElse, type PopupStyleHelpers, type MapSourceLayersSpecification } from '../common';
 
 export const enrrMobilisablesParkingsLayerColor = '#d0643e';
 export const enrrMobilisablesParkingsLayerOpacity = 0.7;
@@ -22,6 +22,7 @@ export const enrrMobilisablesParkingsLayersSpec = [
           'fill-opacity': enrrMobilisablesParkingsLayerOpacity,
         },
         isVisible: (config) => config.enrrMobilisablesSolaireThermique.show && config.enrrMobilisablesSolaireThermique.showParkings,
+        popup: Popup,
       },
       {
         id: 'enrrMobilisables-parkings-contour',
@@ -36,3 +37,22 @@ export const enrrMobilisablesParkingsLayersSpec = [
     ],
   },
 ] as const satisfies ReadonlyArray<MapSourceLayersSpecification>;
+
+type SolaireThermiqueParking = {
+  GmlID: string;
+  TYPE: string;
+  'db_gd5kj.hsu_pnjyu.parkings_sup500m2_parkings_sup500m2.fid': number;
+  st_area_shape_: number;
+  st_length_shape_: number;
+  surfm2: number;
+};
+
+function Popup(parking: SolaireThermiqueParking, { Property, Title }: PopupStyleHelpers) {
+  return (
+    <>
+      <Title>Parking</Title>
+      <Property label="Surface" value={parking.surfm2} unit="m²" />
+      <Property label="Source" value="Cerema" />
+    </>
+  );
+}
