+++ src/components/Map/components/layers/reseauxDeFroid.tsx
@@ -1,4 +1,9 @@
-import { ifHoverElse, type MapSourceLayersSpecification } from './common';
+import Button from '@/components/ui/Button';
+import { type NetworkSummary } from '@/types/Summary/Network';
+import { isDefined } from '@/utils/core';
+import { prettyFormatNumber } from '@/utils/strings';
+
+import { ifHoverElse, type PopupStyleHelpers, type MapSourceLayersSpecification } from './common';
 import { buildFiltreGestionnaire, buildFiltreIdentifiantReseau } from './filters';
 
 export const reseauxDeFroidColor = '#0094FF';
@@ -32,6 +37,7 @@ export const reseauxDeFroidLayersSpec = [
           ...buildFiltreIdentifiantReseau(config.filtreIdentifiantReseau),
         ],
         isVisible: (config) => config.reseauxDeFroid,
+        popup: Popup,
       },
       {
         id: 'reseauxDeFroid-sans-trace',
@@ -50,7 +56,33 @@ export const reseauxDeFroidLayersSpec = [
           ...buildFiltreIdentifiantReseau(config.filtreIdentifiantReseau),
         ],
         isVisible: (config) => config.reseauxDeFroid,
+        popup: Popup,
       },
     ],
   },
 ] as const satisfies ReadonlyArray<MapSourceLayersSpecification>;
+
+function Popup(reseauDeFroid: NetworkSummary, { Property, Title }: PopupStyleHelpers) {
+  return (
+    <>
+      <Title>{reseauDeFroid.nom_reseau ?? 'Réseau de froid'}</Title>
+      <Property label="Identifiant" value={reseauDeFroid['Identifiant reseau']} />
+      <Property label="Gestionnaire" value={reseauDeFroid.Gestionnaire} />
+      <Property
+        label="Contenu CO2 ACV"
+        value={reseauDeFroid['contenu CO2 ACV']}
+        formatter={(value) => (isDefined(value) ? `${prettyFormatNumber(value * 1000)} g/kWh` : 'Non connu')}
+      />
+      {reseauDeFroid['Identifiant reseau'] && (
+        <Button
+          priority="secondary"
+          full
+          iconId="fr-icon-eye-line"
+          linkProps={{ href: `/reseaux/${reseauDeFroid['Identifiant reseau']}`, target: '_blank', rel: 'noopener noreferrer' }}
+        >
+          Voir la fiche du réseau
+        </Button>
+      )}
+    </>
+  );
+}
