+++ src/components/Map/components/layers/typeChauffageBatimentsCollectifs.tsx
@@ -5,7 +5,7 @@ import { type EnergySummary } from '@/types/Summary/Energy';
 import { deepMergeObjects } from '@/utils/core';
 import { ObjectEntries } from '@/utils/typescript';
 
-import { ifHoverElse, intermediateTileLayersMinZoom, type MapSourceLayersSpecification } from './common';
+import { ifHoverElse, intermediateTileLayersMinZoom, type PopupStyleHelpers, type MapSourceLayersSpecification } from './common';
 import { type MapLayerSpecification } from '../../map-layers';
 
 export const minIconSize = 12;
@@ -155,6 +155,7 @@ export const typeChauffageBatimentsCollectifsLayersSpec = [
           ];
         },
         isVisible: (config) => config.batimentsFioulCollectif.show || config.batimentsGazCollectif.show,
+        popup: Popup,
       }),
     ],
   },
@@ -222,6 +223,22 @@ function buildLayerAndHoverLayer<LayerId extends string>(
         ],
       },
       unselectable: true,
+      popup: undefined, // overwride
     },
   ] as const satisfies ReadonlyArray<MapLayerSpecification>;
+}
+
+function Popup(caracteristiqueBatiment: EnergySummary, { Property, Title }: PopupStyleHelpers) {
+  return (
+    <>
+      <Title>{caracteristiqueBatiment.addr_label}</Title>
+      <Property label="Année de construction" value={caracteristiqueBatiment.annee_construction} />
+      <Property label="Usage" value={caracteristiqueBatiment.type_usage} />
+      <Property label="Nombre de logements" value={caracteristiqueBatiment.nb_logements} />
+      <Property label="Chauffage actuel" value={caracteristiqueBatiment.energie_utilisee} />
+      <Property label="Mode de chauffage" value={caracteristiqueBatiment.type_chauffage} />
+      <Property label="DPE consommations énergétiques" value={caracteristiqueBatiment.dpe_energie} />
+      <Property label="DPE émissions de gaz à effet de serre" value={caracteristiqueBatiment.dpe_ges} />
+    </>
+  );
 }
