+++ src/types/MapComponentsInfos.ts
@@ -1,19 +1,3 @@
-import { type DynamicPopupContentType } from '@/components/Map/components/DynamicMapPopupContent';
-
-export type MapPopupInfos = {
-  latitude: number;
-  longitude: number;
-  content: DynamicPopupContentType | { [x: string]: any };
-};
-
-export enum MapPopupType {
-  DEFAULT = 'default',
-  ENGIE = 'engie',
-  DALKIA = 'dalkia',
-  VIASEVA = 'viaseva',
-  IDEX = 'idex',
-}
-
 export type MapMarkerInfos = {
   id: string;
   latitude: number;
