--- src/types/layers/CommuneFortPotentielPourCreationReseauxChaleur.ts
@@ -1,10 +1,0 @@
-export type CommuneFortPotentielPourCreationReseauxChaleur = {
-  id: number;
-  insee_dep: string;
-  insee_com: string;
-  nom: string;
-  population: number;
-  zones_fort_potentiel_total: number;
-  zones_fort_potentiel_chauf_mwh: number;
-  zones_fort_potentiel_ecs_mwh: number;
-};
